OPENEUROPA-2251: Improvements after the review.
